Update Pairs trading
Script takes stock data allowing it to be visualised with some basic indicators applied to help basic analysis. Statistical tests then applied to stocks data if desired in order to determine suitability for pairs trading. If suitable then a back test is performed with feature engineering to fit the best parameters for creating trade signals.